Fix compil issue

+++ src/GeneticAlgorithm.cpp
@@ -20,12 +20,13 @@ GeneticAlgorithm::GeneticAlgorithm(const Config &config, Simulator &simulator,
     }
 }
 
-bool GeneticAlgorithm::canExecuteProcess(const Process* process, const std::map<std::string, int>& stocks) const {
+bool GeneticAlgorithm::canExecuteProcess(
+    const Process *process, const std::map<std::string, int> &stocks) const {
     if (!process) {
         return false;
     }
 
-    for (const auto& [resource, quantity] : process->inputs) {
+    for (const auto &[resource, quantity] : process->inputs) {
         auto stockIt = stocks.find(resource);
         if (stockIt == stocks.end() || stockIt->second < quantity) {
             return false;
@@ -34,18 +35,19 @@ bool GeneticAlgorithm::canExecuteProcess(const Process* process, const std::map<
     return true;
 }
 
-void GeneticAlgorithm::updateStocksAfterProcess(const Process* process, std::map<std::string, int>& stocks) const {
+void GeneticAlgorithm::updateStocksAfterProcess(
+    const Process *process, std::map<std::string, int> &stocks) const {
     if (!process) {
         return;
     }
-    
+
     // Consume inputs
-    for (const auto& [resource, quantity] : process->inputs) {
+    for (const auto &[resource, quantity] : process->inputs) {
         stocks[resource] -= quantity;
     }
-    
+
     // Add outputs
-    for (const auto& [resource, quantity] : process->outputs) {
+    for (const auto &[resource, quantity] : process->outputs) {
         stocks[resource] += quantity;
     }
 }
@@ -53,57 +55,61 @@ void GeneticAlgorithm::updateStocksAfterProcess(const Process* process, std::map
 Individual GeneticAlgorithm::createSmartIndividual() {
     // Initialize stocks from config
     map<string, int> currentStocks;
-    for (const auto& stock : config.getStocks()) {
+    for (const auto &stock : config.getStocks()) {
         currentStocks[stock.name] = stock.quantity;
     }
-    
+
     // Create a mapping of process names to process pointers for quick lookup
-    unordered_map<string, const Process*> processMap;
-    for (const auto& process : config.getProcesses()) {
+    unordered_map<string, const Process *> processMap;
+    for (const auto &process : config.getProcesses()) {
         processMap[process.name] = &process;
     }
-    
+
     // Create sequence
     vector<string> sequence;
     int sequenceLength = 0;
-    int maxAttempts = maxSequenceLength * 2; // To avoid potential infinite loops
+    int maxAttempts =
+        maxSequenceLength * 2; // To avoid potential infinite loops
     int attempts = 0;
-    
+
     while (sequenceLength < maxSequenceLength && attempts < maxAttempts) {
         attempts++;
-        
+
         // Find executable processes
-        vector<const Process*> executableProcesses;
-        for (const auto& [name, proc] : processMap) {
+        vector<const Process *> executableProcesses;
+        for (const auto &[name, proc] : processMap) {
             if (canExecuteProcess(proc, currentStocks)) {
                 executableProcesses.push_back(proc);
             }
         }
-        
+
         if (executableProcesses.empty()) {
             // No executable processes found, break the loop
             break;
         }
-        
-        // Select a random executable process with preference for processes leading to optimization goals
-        uniform_int_distribution<size_t> dist(0, executableProcesses.size() - 1);
-        const Process* selectedProcess = executableProcesses[dist(randomGenerator)];
+
+        // Select a random executable process with preference for processes
+        // leading to optimization goals
+        uniform_int_distribution<size_t> dist(0,
+                                              executableProcesses.size() - 1);
+        const Process *selectedProcess =
+            executableProcesses[dist(randomGenerator)];
         sequence.push_back(selectedProcess->name);
         sequenceLength++;
-        
+
         // Update stocks
         updateStocksAfterProcess(selectedProcess, currentStocks);
     }
-    
+
     // If sequence is too short, pad it with random processes
     // This helps maintain genetic diversity
-    if (sequence.size() < minSequenceLength) {
+    if (sequence.size() < static_cast<size_t>(minSequenceLength)) {
         uniform_int_distribution<size_t> procDist(0, processNames.size() - 1);
-        while (sequence.size() < minSequenceLength) {
+        while (sequence.size() < static_cast<size_t>(minSequenceLength)) {
             sequence.push_back(processNames[procDist(randomGenerator)]);
         }
     }
-    
+
     return Individual(sequence);
 }
 
@@ -130,17 +136,18 @@ void GeneticAlgorithm::initializePopulation() {
     // Create a mix of smart and random individuals
     // Smart individuals will guide the search, random ones maintain diversity
     int smartCount = populationSize * 0.8; // 80% smart, 20% random
-    
+
     for (int i = 0; i < smartCount; ++i) {
         population.push_back(createSmartIndividual());
     }
-    
+
     for (int i = smartCount; i < populationSize; ++i) {
         population.push_back(createRandomIndividual());
     }
 
-    cout << "Population initialized with " << populationSize << " individuals: " 
-         << smartCount << " smart, " << (populationSize - smartCount) << " random." << endl;
+    cout << "Population initialized with " << populationSize
+         << " individuals: " << smartCount << " smart, "
+         << (populationSize - smartCount) << " random." << endl;
 }
 
 void GeneticAlgorithm::evaluatePopulation() {
